+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- DPS can now optionally add summon damage to the owner's DPS row with the new `Include Summons in Owner DPS` checkbox.
 
 ## [1.7.0] - 2026-04-12
 ### Added